pyproject.toml
[project]
name = "py-experimenter-dashboard"
version = "0.0.1"
description = "This is a dashboard that can be connected to PyExperimenter setups."
authors = [{ name = "Marcel Wever", email = "m.wever@ai.uni-hannover.de" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "aiomysql>=0.2",
    "aiosqlite>=0.20",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    # Chat / code-execution feature
    "httpx>=0.27",
    "pandas>=2.0",
    "matplotlib>=3.8",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.scripts]
py-experimenter-dashboard = "py_experimenter_db.cli:main"

[project.urls]
Homepage = "https://mwever.github.io/py-experimenter-dashboard/"
Repository = "https://github.com/mwever/py-experimenter-dashboard"
Documentation = "https://mwever.github.io/py-experimenter-dashboard/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.1",
    "tox-uv>=1.29.0",
    "deptry>=0.24.0",
    "mypy>=1.19.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["py_experimenter_db"]
include = ["py_experimenter_db/dashboard/templates/**"]

[tool.mypy]
files = ["py_experimenter_db"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["py_experimenter_db"]

